audio-recording-sngle-user/backend/app/requirements.txt
# api framework
fastapi==0.79.0
uvicorn==0.18.2
PyYAML==6.0
python-multipart==0.0.5

# audio
MoviePy==1.0.3
pysoundfile==0.9.0.post1
librosa==0.9.2
pandas==1.4.4
# db orm
sqlalchemy==1.4.39
sqlalchemy-utils==0.38.3
psycopg2==2.9.3

# migrations and docker workflow
wait-for-it==2.2.1
alembic==1.8.1
